Move report_wiring_{error,message} to logging.py
