+++ magma/array.py
@@ -7,7 +7,7 @@ from .compatibility import IntegerTypes
 from .bit import VCC, GND, Bit
 from .bitutils import seq2int
 from .debug import debug_wire, get_callee_frame_info
-from .port import report_wiring_error
+from .logging import report_wiring_error
 
 
 class ArrayMeta(ABCMeta, Kind):
