+++ magma/port.py
@@ -1,5 +1,5 @@
 from .config import get_debug_mode
-from .logging import error, warning, get_source_line
+from .logging import error, warning, get_source_line, report_wiring_error
 from .backend.util import make_relative
 from .t import Direction
 
@@ -12,29 +12,6 @@ __all__ += ['Port']
 INPUT = Direction.In
 OUTPUT = Direction.Out
 INOUT = Direction.InOut
-
-
-def _report_wiring_messgae(fn, message, debug_info):
-    if not debug_info:
-        error(message)
-        return
-    file = debug_info[0]
-    line = debug_info[1]
-    message = f"\033[1m{make_relative(file)}:{line}: {message}"
-    fn(message, include_wire_traceback=True)
-    try:
-        fn(get_source_line(file, line))
-    except FileNotFoundError:
-        fn(f"    Could not file file {file}")
-
-
-def report_wiring_error(message, debug_info):
-    _report_wiring_messgae(error, message, debug_info)
-
-
-def report_wiring_warning(message, debug_info):
-    # TODO(rsetaluri): Include wire traceback support.
-    _report_wiring_messgae(warning, message, debug_info)
 
 
 def flip(direction):
